search trong khoảng

+++ src/main/java/com/example/loverbe/model/User.java
@@ -63,7 +63,7 @@ public class User {
     )
     private Set<Role> roles;
 
-    private Integer viewCounter;
+    private Long viewCounter;
 
     private Integer rentedCounter;
